+++ fixtures/kaocha-demo/demo/test.clj
@@ -22,3 +22,8 @@
 (deftest ^:kaocha/skip skip-test
   (println "this test does not run.")
   (is false))
+
+(deftest ^:kaocha/pending pending-test
+  (println "this test does not run, but is marked as 'skipped'.")
+  (is (= {:foo 1} {:foo 1})
+      "This test would pass, if it wasn't skipped"))